Updated methods for sending, updating, deleting

+++ backend/src/main/java/ru/zan/Pulsometer/repositories/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.zan.Pulsometer.models.User;
 
 @Repository
-public interface UserRepository extends R2dbcRepository<User, Long> {
+public interface UserRepository extends R2dbcRepository<User, Integer> {
 
 }
